uses resq for static long because 64 bit can and it will always be right, also fixes a potential bug with 64 bit operations that requried jumps

+++ Compiler/a5.tst/cs444/acceptance/AsmAndLinkCallback.java
@@ -37,6 +37,8 @@ import cs444.codegen.Platform;
         public boolean beforeCompile(final File _) throws IOException {
             final File outputDir = new File(Compiler.OUTPUT_DIRECTORY);
             for(final File platformFolder : outputDir.listFiles()) {
+                //some computers (my mac) make files like .DS_Store...
+                if (platformFolder.isFile()) continue;
                 deleteNonRuntime(platformFolder);
             }
             return true;
